ebpf_domain_t; no abs_transformer_t
