+++ src/crab/adapt_sgraph.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "crab/vec.hpp"
+#include "crab/debug.hpp"
+#include "crab/types.hpp"
 // Adaptive sparse-set based weighted graph implementation
 
 #pragma GCC diagnostic push
