+++ src/crab/ebpf_domain.cpp
@@ -1,4 +1,4 @@
-#include "crab/array_expansion.hpp"
+#include "crab/ebpf_domain.hpp"
 
 namespace crab {
 namespace domains {
